requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
pytest==8.1.1
python-dotenv==1.0.1
dynaconf==3.2.5
pylint==3.1.0